I set up the android project
We will tentatively call the app "Breathe".

I set up the project with an empty activity. I set the package name to com.ybeltagy.breathe. I set the minimum SDK API as 22 (Android Lollipop 5.1).

I added the .gitignore file before the rest of the project to ensure the specified files are ignored.