ClimaCell API calls
I implemented the logic for making HTTP requests to get weather information from the ClimaCell API.

I also wrote unit tests to test the feature in Seattle, Spokane, Miami, and a random location in the US.

I need to write test cases to test retrieving historical information and clean up the code.

I have yet to connect this logic with a Broadcast receiver. The next steps would be to either implement the logic that gets the pin data or implement a Broadcast receiver and test the logic for the Inhaler electronics.