Smart pin+breathe app integration (#23)
- Integrated the smart-pin and the app
- Modified the smart-pin to send all data in one object
- Add a foreground service that is running in the background of the phone all the time
- Added a repository level .gitignore

The phone autoconnects to smart-pin now. This happens even after reboot or restarting the Bluetooth adapter.

The Wearable Sensor requires 'Encryption' now. I added this because Android mostly abstracts the bonding process from the application developer. One of the ways to enforce bonding is for the peripheral device itself to require it.