wearable: fixed dht11 data issue
The dht11 data would not update when the wearable was connected to a power socket rather than a PC.

Updated the FUS/BLE stack and made this commit to fix this issue. Even after updating the FUS/BLE stack, the issue persisted. But when I modified the dht11's timer init, the issue was solved. However it is hard to confirm that the commit was the solution because the FUS/BLE stack may have been the issue and the MCU might have just acted weird that first time.